Implement nqp::isint and nqp::isnum for JS
The isnum implementation overlaps with isint a bit; since JavaScript
only has a number type, we can't really tell if a number was intended
to be used as an integer or a real number.  This will make a test in the
NQP test suite fail, but I don't know if it will interfere with the
operations of NQP that much.